update more unit tests to use mock querier instead of mock contracts
