+++ contracts/voting-verifier/tests/tests.rs
@@ -67,128 +67,6 @@ fn messages(len: u64) -> Vec<Message> {
 }
 
 #[test]
-fn should_query_message_statuses() {
-    let mut fixture = setup();
-
-    let messages = messages(10);
-    let msg = msg::ExecuteMsg::VerifyMessages {
-        messages: messages.clone(),
-    };
-
-    let res = fixture
-        .voting_verifier
-        .execute(&mut fixture.app, Addr::unchecked(SENDER), &msg)
-        .unwrap();
-
-    let reply: msg::VerifyMessagesResponse = from_binary(&res.data.unwrap()).unwrap();
-
-    assert_eq!(reply.verification_statuses.len(), messages.len());
-    assert_eq!(
-        reply.verification_statuses,
-        messages
-            .iter()
-            .map(|message| (message.cc_id.clone(), VerificationStatus::None))
-            .collect::<Vec<(_, _)>>()
-    );
-
-    let statuses: Vec<(CrossChainId, VerificationStatus)> = fixture.voting_verifier.query(
-        &fixture.app,
-        &msg::QueryMsg::GetMessagesStatus {
-            messages: messages.clone(),
-        },
-    );
-    assert_eq!(
-        statuses,
-        messages
-            .iter()
-            .map(|message| (message.cc_id.clone(), VerificationStatus::InProgress))
-            .collect::<Vec<(_, _)>>()
-    );
-
-    let msg: msg::ExecuteMsg = msg::ExecuteMsg::Vote {
-        poll_id: Uint64::one().into(),
-        votes: (0..messages.len())
-            .map(|i| {
-                if i % 2 == 0 {
-                    Vote::SucceededOnChain
-                } else {
-                    Vote::NotFound
-                }
-            })
-            .collect::<Vec<_>>(),
-    };
-
-    fixture
-        .voting_verifier
-        .execute(&mut fixture.app, Addr::unchecked("addr1"), &msg)
-        .unwrap();
-
-    fixture
-        .voting_verifier
-        .execute(&mut fixture.app, Addr::unchecked("addr2"), &msg)
-        .unwrap();
-
-    fixture
-        .app
-        .update_block(|block| block.height += POLL_BLOCK_EXPIRY);
-
-    let msg: msg::ExecuteMsg = msg::ExecuteMsg::EndPoll {
-        poll_id: Uint64::one().into(),
-    };
-
-    fixture
-        .voting_verifier
-        .execute(&mut fixture.app, Addr::unchecked(SENDER), &msg)
-        .unwrap();
-
-    let statuses: Vec<(CrossChainId, VerificationStatus)> = fixture.voting_verifier.query(
-        &fixture.app,
-        &msg::QueryMsg::GetMessagesStatus {
-            messages: messages.clone(),
-        },
-    );
-
-    assert_eq!(
-        statuses,
-        messages
-            .iter()
-            .enumerate()
-            .map(|(i, message)| (
-                message.cc_id.clone(),
-                if i % 2 == 0 {
-                    VerificationStatus::SucceededOnChain
-                } else {
-                    VerificationStatus::NotFound
-                }
-            ))
-            .collect::<Vec<(_, _)>>()
-    );
-}
-
-#[test]
-fn should_start_worker_set_confirmation() {
-    let mut fixture = setup();
-
-    let operators = Operators::new(vec![(vec![0, 1, 0, 1].into(), 1u64.into())], 1u64.into());
-    let msg = msg::ExecuteMsg::VerifyWorkerSet {
-        message_id: message_id("id", 0),
-        new_operators: operators.clone(),
-    };
-    let res = fixture
-        .voting_verifier
-        .execute(&mut fixture.app, Addr::unchecked(SENDER), &msg);
-    assert!(res.is_ok());
-
-    let res: VerificationStatus = fixture.voting_verifier.query(
-        &fixture.app,
-        &msg::QueryMsg::GetWorkerSetStatus {
-            new_operators: operators.clone(),
-        },
-    );
-    assert_eq!(res, VerificationStatus::InProgress);
-}
-
-#[test]
 fn should_confirm_worker_set() {
     let mut fixture = setup();
